requirements.txt
classes==0.4.1
colorama==0.4.6
msal==1.33.0
numpy==2.3.1
pandas==2.3.1
portalocker==3.2.0
python-dotenv==1.1.1
PyYAML==6.0.2
Requests==2.32.4
openpyxl>=3.1.0